Repeats searching: K-mer count, ClustalO, Repeats Exploring 3 possibilities of searching for repeating sequences
kmercount.py:
Tool outputs counts of k-meres. This is fast but not very usefull

allignment.sh:
Tool extracts matches found by ModDotPlot and alligns them using ClustalO.
This is slow as heck and throws errors on my PC.

repeats.py:
Tool creates graph of k-meres with scores proportional to number of occurences
count of each (k+1)-mere as weights of edges. Then it uses this graph as finite
automaton allowing insertions and gaps at a cost. The idea behind this tool is
something between Smith-Waterman algorithm and Seqtk telo.
This tool is slow but seems to work pretty well.