repeats.py: Batching of values
Added batching for values.

When searching for possible gaps, this commit add batching. In contrast to the previous implementation
in which the script searched for all possible sequences with any-length gap, this implements it as an
iterative process, in which score is reduced down to _flood (minimal value for the iteration, either
`BasePenalty * position` or normalized `max_value`). The next states are stored in sets according to
their new value, batched by decrements of GapPenalty. By this the script stays in (hopefully) linear time
to number of states needed to go through.